Customers Entering and Leaving basics
Added customers walking to the counter, and leaving the counter once an order is complete.
Renamed NewOrder.cs to CustomerOrder.cs

Bug fixes:
Fixed bug where deluxe burgers were not properly removed from the player inventory

Bugs added:
The customer order is taken when the customer interacts with any trigger. This means the order is taken when the customer walks up and hits the counter.
The customer stops walking when it hits any trigger, instead of the customer stops walking when it hits the counter. This is not a problem right now, but it should be fixed in case the customer collides with anything else before it reaches the counter.