Customer Order and Movement Bugfixes
Customer now only stops when it's collision has the "Counter" tag. (This means that the customer will fly across the map unless the tag "Counter" is added to the location of the counter.)

Bug fixes:
The customer only takes the order from the player with the "Player" tag instead of anything it collides with. (This means that the player now requires the "Player" tag in order to generate or receive parts of an order.)