Fixed Soda Can Mechanic
Bug Fixes:
Fixed the collision being bigger than it should be. (The player could previously interact with the can on the floor. This is fixed so the player can only interact with the floor when the drink is spilled.)

Stopped the player from re-opening the drink (Previously the can was reset back to stage 0 after the can is spilled in stage 0, allowing the can to be reopened. This is fixed by setting the can to an unused stage 3 after spilled.)